feat: separates list contacts page in multiple components
