+++ frontend/src/pages/ListContactsPage/components/SearchNotFound/styles.tsx
@@ -1,0 +1,15 @@
+import { styled } from "../../../../../styled-system/jsx";
+
+export const SearchNotFoundContainer = styled("div", {
+  base: {
+    display: "flex",
+    flexDirection: "column",
+    alignItems: "center",
+    w: "3/4",
+    mt: "12",
+    mx: "auto",
+    gap: "2",
+    textAlign: "center",
+    color: "primary.light",
+  },
+});
